exchange/requirements.txt
python-dotenv==1.0.0
pika==1.3.2